Upgrade*Command - Generate consistent 128-bits of entropy in `name`
On OSX, the default behavior of rand() only produces 31 bits of entropy.
MD5 hashing is 128-bits, so we've left a lot of gaps in the 128-bit space.
(Effectively: 93 vs 128 bits.) The PHP docs say that rand() on Windows gives
even less entropy.

This revision ensures a more consistent amount of entropy.